Creating adhoc TLS certificates
If use_tls is set to true, the role activates all the features
needed to have Quay listening for TLS connections. However, the
resulting registry instance still waits for the user to place the
TLS artifacts (key and certificates) under the configuration
directory.

This change creates a self-signed certificate and adds it to the
ca-trust repository in the ansible controller (jumpbox) so later
accessess to the registry over HTTPS will work seamlessly.

- Updated documentation
- Processed comments from raperez
- Replaced use of host_public_ip by host_fqdn when use_tls=true

Change-Id: I74ed80cab8671b22ff3ed072f79b44575a22411d